Preparing for upcoming `biblatex-unified` (CTAN)
Our biblatex bibliography style will soon be submitted to CTAN. The new
name will be `biblatex-unified` and the style files will be
`unified.{bbx,cbx}`. This revision of `sp.cls` checks whether the new
style files are present and if so, uses them. Otherwise, it uses the old
ones. And if all else fails, as before, it reverts to BibTeX.