hyperxmp now loaded after hyperref
As of hyperxmp v5.12, dated 2023/09/10: “To use hyperxmp, merely put a \usepackage{hyperxmp} in your document’s pream- ble. That line can appear anywhere after the \usepackage{hyperref} but before hyperref’s pdf options are specified with \hypersetup.”

We should see whether this new version lets us construct the xmp data in a way that works both in Acrobat and in Preview.