fix: restore label formatting from prior semprag version
This commit restores label formatting that Chris Brown had merged in
a4cd5f17f1d31a0072927f379529284a160d21f7, but is not present in the Alan
Munn version of the package. I'm not entirely sure where these
originally come from or what "upstream source repo" refers to. There's a
whole bunch of other adjustments in that commit that I'm setting aside
for now.

The basic xlist case of this was tested in a couple of papers in March,
and I have tested this full set of changes in place in those papers. I
think that most of the relevant label types aren't ones we routinely
use.)